Created ReservationForm.js and implemented change to NewReservation.js
